finish routes, needs testing
